refactor(security): keep workspace filesystem checks in executor
