+++ packages/core/src/workspace-paths.ts
@@ -1,4 +1,3 @@
-import { lstat, realpath } from 'node:fs/promises';
 import path from 'node:path';
 
 function assertRelativeIdentity(value: string, label: string): void {
@@ -29,54 +28,19 @@ export function deriveManagedBranchPath(
   return candidate;
 }
 
-async function nearestExistingPath(candidate: string, floor: string): Promise<string> {
-  let current = candidate;
-  while (true) {
-    try {
-      await lstat(current);
-      return current;
-    } catch (error) {
-      if ((error as NodeJS.ErrnoException).code !== 'ENOENT') throw error;
-    }
-    if (current === floor) return floor;
-    const parent = path.dirname(current);
-    if (parent === current) return current;
-    current = parent;
-  }
-}
-
 /**
- * Revalidate a stored workspace at a filesystem sink. Exact derivation catches
- * stale/caller-controlled rows; canonical ancestor comparison catches symlink
- * escapes before a missing workspace is created.
+ * Revalidate a stored workspace against trusted identity. Filesystem-owning
+ * execution substrates must additionally canonicalize existing ancestors.
  */
-export async function assertManagedBranchPath(options: {
+export function assertManagedBranchPath(options: {
   root: string;
   repoSlug: string;
   branchName: string;
   storedPath: string;
-}): Promise<string> {
+}): string {
   const expected = deriveManagedBranchPath(options.root, options.repoSlug, options.branchName);
   if (path.resolve(options.storedPath) !== expected || options.storedPath !== expected) {
     throw new Error('Stored branch workspace does not match the trusted storage layout');
-  }
-
-  const root = path.resolve(options.root);
-  const existingRoot = await nearestExistingPath(root, path.parse(root).root);
-  // If the tenant root itself has not been created yet, there cannot be a
-  // symlink below it. The lifecycle identity will create the derived path.
-  if (existingRoot !== root) return expected;
-  const existingCandidate = await nearestExistingPath(expected, root);
-  const [canonicalRootBase, canonicalCandidate] = await Promise.all([
-    realpath(existingRoot),
-    realpath(existingCandidate),
-  ]);
-  const canonicalRoot = path.join(canonicalRootBase, path.relative(existingRoot, root));
-  if (
-    canonicalCandidate !== canonicalRoot &&
-    !canonicalCandidate.startsWith(`${canonicalRoot}${path.sep}`)
-  ) {
-    throw new Error('Branch workspace resolves outside the managed root');
   }
   return expected;
 }
